tests/It/requirements.txt
mysql-connector-python==8.0.24
mysqlclient==2.0.3